Make Footer a component
