+++ client/src/Pages/HomePage/HomePage.jsx
@@ -16,19 +16,7 @@ const data = [
   { rank: 2, name: 'Joshua', score: 4500 },
   { rank: 3, name: 'JS', score: 4200 },
 ];
-export function Footer() {
-  return (
-    <footer style={{ backgroundColor: '#1a1a1a', color: 'white', padding: '20px 0' }}>
-      <Container size="md" style={{ textAlign: 'center' }}>
-        <Anchor href="https://www.thecoderschool.com/locations/montgomery/" >The Coder School Montgomery</Anchor>
-        <Text>Got a question? Email us at <Anchor href="mailto:montgomery@thecoderschool.com" >montgomery@thecoderschool.com</Anchor></Text>
 
-        <Text size="sm">&copy; 2014-2022, the Coder School San Francisco, LLC. All Rights, Alright?
-          Don't forget our<Anchor href="https://www.thecoderschool.com/privacy/"> Privacy Policy</Anchor></Text>
-      </Container>
-    </footer>
-  )
-}
 export function Leaderboard() {
   const rows = data1.map((data1) => (
